Update ghcjs-setup.sh to include `cabal-js`
I wanted to use this script to install `ghcjs` while also allowing me to use the `cabal` that was compiled with `ghcjs` support. It took me a while to figure out how to do this and I think this would be beneficial to others wanting to build `GHCJS` projects via `cabal-js`.

Normally I wouldn't make an amendment to this build script for such a trivial change, but it really took me a while to figure out why `haskellPackages_ghcjs.cabalInstall` wouldn't compile.